новое меню, звук при перелистывании
настройки звка (вкл/выкл)
перенес скрипты из /new
новое меню
добавил звук листания карты
etc
