+++ Assets/Scripts/CardMachine.cs.meta
