+++ Assets/Scripts/DescriptionUpdater.cs
